fix(stars): close final menu and monitoring races
