+++ src/index.ts
@@ -78,6 +78,7 @@ import {
   setBotInstance,
   verifyPaymentByTxid,
 } from './services/btc-payment';
+import { isStarsMode } from './services/stars-payment';
 import { getStatusText } from './services/admin-stats';
 import { scheduleDatabaseBackups } from './services/backup-service';
 import { handleUpgrade } from 'controllers/upgrade';
@@ -164,6 +165,7 @@ async function updateUserCommands(
   isAdmin: boolean,
   isPremium: boolean,
 ) {
+  if (isStarsMode()) return;
   const locale = ctx.from?.language_code || 'en';
   const commands = [...getBaseCommands(locale)];
   if (isPremium || isAdmin) {
@@ -1300,11 +1302,13 @@ async function startApp() {
   startMonitorLoop();
   resumePendingChecks();
   scheduleDatabaseBackups();
-  await bot.telegram.setMyCommands(getBaseCommands('en'));
-  await bot.telegram.setMyCommands(
-    [...getBaseCommands('en'), ...getPremiumCommands('en'), ...getAdminCommands('en')],
-    { scope: { type: 'chat', chat_id: BOT_ADMIN_ID } }
-  );
+  if (!isStarsMode()) {
+    await bot.telegram.setMyCommands(getBaseCommands('en'));
+    await bot.telegram.setMyCommands(
+      [...getBaseCommands('en'), ...getPremiumCommands('en'), ...getAdminCommands('en')],
+      { scope: { type: 'chat', chat_id: BOT_ADMIN_ID } }
+    );
+  }
   bot.launch({ dropPendingUpdates: true }).then(() => {
     console.log('✅ Telegram bot started successfully and is ready for commands.');
   });
